move deploy and async_deploy functions to new utils file
